move ME Cypress helpers to other file
